refactor(weex): Adjust the weex platform entry file to fit the new weex runtime (#6620)
* trim trailing whitespace

* revert(weex): remove the new Function hack for V8, as Weex uses JSC now

Remove the `callFunctionNative` method and `compileBundle`, which is provided by modified V8. In
order to maintain the consistency of the js engine, Weex also use JSC on Android. The legacy hack
for V8 engine should be removed.

* refactor(weex): move module and component apis to weex runtime

Modules and components should be registered in weex runtime, not the specific framework. The
`registerModules`, `registerComponents`, `weex.supports` and `weex.requireModule` api is moved to
weex runtime, which is in the "apache/incubator-weex" repo.

* test(weex): refactor the test cases to fit the new weex-js-runtime

* fix(weex): use document instead of renderer to create Element